don't raise DeadActorErrors when publishing from a worker
It is normal for Envoy to be shutdown while processing a message,
and after the message is processed the worker would attempt to
publish a notification to say it had succeeded/failed. At this point the
Notifications actor would already be dead, and Celluloid would
raise a DeadActorError. This is an acceptable error and the worker
should not Airbrake or log out an error.